requirements.txt
setuptools==3.8.1
fastapi==0.82.0
uvicorn==0.11.3
gunicorn
werkzeug==0.14.1
requests
pytz
python-dateutil
psycopg2-binary
werkzeug==0.14.1
PyJWT
python-decouple